# api/gen/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  description: |
    An example OpenAPI specification for the Little PM application, referencing path parameters from components.
  title: Little PM API
  version: 1.0.0
servers:
- description: Production server
  url: https://app.little.pm
...
paths:
  /auth:
    post:
      description: Authenticate a user and return a token or session details.
      operationId: authenticateUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserRegistration'
        required: true
      responses:
        "200":
          description: OK
          headers:
            Set-Cookie:
              $ref: '#/components/headers/Set-JWT'
            Location:
              $ref: '#/components/headers/HomeRedirect'
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized login
      summary: User authentication
      tags:
      - User Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: User Service
...
  /finances/{projectUUID}:
    get:
      description: Retrieve budget or finance info for a specific project.
      operationId: getFinances
      parameters:
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
          description: Financial details found
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Project not found
      summary: Get finance details
      tags:
      - Finances Service
      x-accepts:
      - application/json
      x-tags:
      - tag: Finances Service
    post:
      description: "Update or create financial info (e.g., budget) for a project."
      operationId: updateFinances
      parameters:
      - description: ID of the project
        explode: false
        in: path
        name: projectUUID
        required: true
        schema:
          $ref: '#/components/schemas/projectUUID'
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetUpdate'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
          description: Finance details updated/created
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Project not found
      summary: Update or create finance details
      tags:
      - Finances Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Finances Service
  /finances/calculate:
    post:
      description: "Given investment data, compute ROI or related metrics."
      operationId: calculateFinances
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InvestmentData'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ROIResult'
          description: Calculation result
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Unauthorized access
      summary: Calculate ROI or financial projections
      tags:
      - Finances Service
      x-content-type: application/json
      x-accepts:
      - application/json
      x-tags:
      - tag: Finances Service
...
components:
  headers:
    Set-JWT:
      explode: false
      schema:
        description: JWT token
        example: jwt=tokenValue; Path=/; HttpOnly; Secure
        type: string
      style: simple
...
    HomeRedirect:
      explode: false
      schema:
        description: Home redirect
        example: /
        type: string
      style: simple
...
  schemas:
...
    projectUUID:
      description: ID of the project
      example: 665c599d-5c8d-4d20-aaab-7ffaba150606
      format: uuid
      type: string
...
    Error:
      description: Error object
      example:
        message: Error message
      properties:
        message:
          description: Error message
          example: Error message
          type: string
      type: object
    UserRegistration:
      example:
        password: password
        name: name
        email: email
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
      type: object
...
    Budget:
      example:
        totalFunds: 0.8008282
        allocated: 6.0274563
      properties:
        totalFunds:
          format: float
          type: number
        allocated:
          format: float
          type: number
      type: object
    BudgetUpdate:
      example:
        totalFunds: 0.8008281904610115
        allocated: 6.027456183070403
      properties:
        totalFunds:
          type: number
        allocated:
          type: number
      type: object
    InvestmentData:
      example:
        initialInvestment: 0.8008281904610115
        expenses: 6.027456183070403
      properties:
        initialInvestment:
          type: number
        expenses:
          type: number
      type: object
    ROIResult:
      example:
        roiPercentage: 0.8008281904610115
        profit: 6.027456183070403
      properties:
        roiPercentage:
          type: number
        profit:
          type: number
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token
      scheme: bearer
      type: http